chore: add envCheck and update ci env
