+++ .circleci/config.yml
@@ -301,12 +301,11 @@ jobs:
           name: Set branch environment
           command: |
             echo 'export REACT_APP_BRANCH=${CIRCLE_BRANCH}' >> $BASH_ENV
-            echo 'export NODE_OPTIONS=--max-old-space-size=4096' >> $BASH_ENV
             echo 'export REACT_APP_PROJECT_VERSION=${CIRCLE_SHA1}' >> $BASH_ENV
       - run:
           name: Check environment variables
           command: |
-            echo REACT_APP_BRANCH=$REACT_APP_BRANCH NODE_OPTIONS=$NODE_OPTIONS
+            echo REACT_APP_BRANCH=$REACT_APP_BRANCH
             echo $REACT_APP_PROJECT_VERSION
       - run:
           command: yarn build
