+++ .env-cmdrc.js
@@ -1,3 +1,7 @@
+// Run a pre-flight check that developer environment setup in compatible way
+const { envCheck } = require('./scripts/envCheck')
+envCheck()
+
 // Depending on node version use different environment variables to fix
 // specific build or run issues
 
@@ -5,11 +9,14 @@ const NODE_VERSION = process.versions.node.split('.')[0]
 
 // Specific settings to use when running anything that requires a webpack compiler
 // Enabled when npm command specifies `env-cmd -e webpack`
-let webpack = {}
+let webpack = {
+  // fix out-of-memory issues
+  NODE_OPTIONS: '--max-old-space-size=4096',
+}
 if (NODE_VERSION > '17') {
   // fix https://github.com/facebook/create-react-app/issues/11708
   // https://github.com/facebook/create-react-app/issues/12431
-  webpack.NODE_OPTIONS = '--openssl-legacy-provider --no-experimental-fetch'
+  webpack.NODE_OPTIONS += ' --openssl-legacy-provider --no-experimental-fetch'
 }
 
 // Specific env to use with react-scripts / create-react-app
